WIP: kernel-netlink: Install a route in table 220 for installed virtual IPs
An additional check on the source IP and the routing table when
adding routes was added with Linux kernel 4.3, which prevents our source
routes in table 220 from getting installed.

WIP: Not sure if this is how we are supposed to do this.  Anyway, it
won't fix installation of the routes on gateways (or in general if we
don't install a virtual IP). I guess we'd could "copy" the route for the
local address to table 220.

 #1189